requirements.txt
selenium==4.15.2
webdriver-manager==4.0.1
beautifulsoup4==4.12.2
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
joblib==1.3.2
gspread==5.12.0
google-auth-oauthlib==1.2.0
google-auth==2.25.2
google-api-python-client==2.106.0
gspread-dataframe==3.3.1
requests==2.31.0
python-dotenv==1.0.0
